Timed budgets: midnight reset option, rolling limit; windows may cross midnight
Mirrors customBlocker 0f0bcef in Mac Vault's engine and editor.

- Schedule: TimeWindow may cross midnight (2300-0100); the part after
  midnight belongs to the day the window starts. Only start == end is invalid.
- BlockGroup: allowedMinutes / resetIntervalHours are Double (a 2.5 h
  interval was truncated to 2), plus resetAtMidnight and rollingLimit.
- UsageBudget (new): the same period / rolling-window rules as the extension.
  reconcileUsage uses it: fixed budgets reset on the anchor or midnight grid;
  rolling groups keep per-minute usage and their timer is the in-window total.
- Hub: clusters share rolling usage per minute (usageBuckets deltas,
  usageBucketsSeed until the first delta), pruned past the longest window.
- Web store / shim carry usageBucketsMs so the editor shows the live rolling
  countdown and the popup's wholesale store write keeps it.
- Editor mirror: the two options, the next-reset / time-returns line, and
  the English placeholder strings (translation pending).

Tests: UsageBudgetTests, ClusterRollingUsageTests, cross-midnight schedule
and fractional import cases.

+++ Sources/MacBlockerScreenTime/DeviceActivityRuleScheduler.swift
@@ -6,14 +6,14 @@ public struct DeviceActivityScheduleRequest: Codable, Equatable, Sendable {
     public var name: String
     public var weekdays: Set<Weekday>
     public var windows: [TimeWindow]
-    public var thresholdMinutes: Int?
+    public var thresholdMinutes: Double?
 
     public init(
         groupID: String,
         name: String,
         weekdays: Set<Weekday>,
         windows: [TimeWindow],
-        thresholdMinutes: Int?
+        thresholdMinutes: Double?
     ) {
         self.groupID = groupID
         self.name = name
